Update main.cpp
Made it give back some info to the user to tell them why its not working. Mostly to do with the exe being in the wrong location or certain files set to read only.

Also added a SCOPE.ini file, the program will check for it and use the values it finds there first and if not will generate a new one and use that instead. This means that the user can mess with any CVARS in any way they like, not just the ones in the default ini file.